fix(sensor-tag-wifi): align sample_int NVS type with console putUShort
Serial console stores values <=65535 via putUShort (NVS type U16), but
loadConfig() read sample_int via getUInt (U32). NVS returns the default
on a type mismatch, so 'set sample_int 600' silently had no effect.

- Change Config::sampleIntervalSec and DEFAULT_SAMPLE_INTERVAL_SEC to
  uint16_t and read with getUShort. U16 max is 18h — ample for a solar
  sensor node.
- Add sample_int to the printKeyValue U16 dispatch alongside mqtt_port
  so 'get sample_int' displays correctly instead of falling through to
  getUChar and showing 0.

+++ firmware/sensor-tag-wifi/include/config.h
@@ -21,7 +21,7 @@ constexpr uint8_t PIN_BATTERY_ADC = 0;
 // Power / Timing Defaults
 // =============================================================================
 
-constexpr uint32_t DEFAULT_SAMPLE_INTERVAL_SEC = 300;   // 5 min
+constexpr uint16_t DEFAULT_SAMPLE_INTERVAL_SEC = 300;   // 5 min (U16 max = 18h)
 constexpr uint8_t  DEFAULT_UPLOAD_EVERY_N      = 1;     // Upload every sample
 constexpr uint16_t WIFI_CONNECT_TIMEOUT_MS     = 10000;
 constexpr uint16_t MQTT_CONNECT_TIMEOUT_MS     = 5000;
